+++ changelog.md
@@ -1,5 +1,18 @@
 # Changelog
 
+
+## 2026-07-27
+
+### Fixed
+- read_until_delimiter uses a buffer correctly and only deserializes the required slice, tests pass now 
+
+### Changed
+- command_to_string requires a flag for if the command will return zero bytes 
+
+### Added
+- framework for networking code testing
+- subtrait of read + write: IOStream. made for allowing generics for duplexes 
+
 ## 2026-07-27
 
 ### Added